Create a encoder-decoder neural network model (#3) -Model is re-implemented with keras framework

+++ boolean_function_generator.py
@@ -10,9 +10,9 @@ def booleanFunctionGenerator(function,dimension):
 
     if ( vector_size < binary_size ):
         print("Invalid function for this dimension.\n")
-        return numpy.zeros( vector_size, dtype=numpy.int8 )
+        return numpy.zeros( vector_size, dtype=numpy.float32 )
     
-    sign_vector = numpy.ones( vector_size, dtype=numpy.int8 )
+    sign_vector = numpy.ones( vector_size, dtype=numpy.float32 )
     for iterator in range (binary_size):
         if ( binary_number[binary_size - iterator - 1] == '1' ):
             sign_vector[iterator] = -1
